refactor: Moved tools to its own file WIP
