+++ internal/text/querier_setup_tools_test.go
@@ -1,0 +1,1 @@
+package text
